Got standard isolation forest R/W working after major refactor. Still a work in progress.
